cla_contact_rx_task: Add EID spoofing detection
To validate if a foreign DTN node pretends to be the receiver itself by
spoofing the receiver's EID an additional check of the sender's EID was
added to the RX task. If the bundle was received via a CLA and the
source EID equals the destination EID the bundle is dropped.

Signed-off-by: Maximilian Nitsch <maximilian.nitsch@d3tn.com>