bundle_processor: Add timeouts for status reports and custody signals
In the case of failed forwarding attempts (and the resulting bundle
re-forwarding), we already have a timeout in case the router task is
currently blocked. However, in the case of status reports and custody
signals (which might as well occur when there is a failed forwarding
attempt), we did not set any timeout. This led to the possibility of
deadlocks between the bundle processor and the router task if such
features were used under high traffic load. See #3.

This also removes the possibility of a bundle memory leak in case the
hop count check fails. Now bundles are always properly deleted if
bundle_forward() encounters an error.

Fixes: #3

Signed-off-by: Felix Walter <felix.walter@d3tn.com>