cla_tcp_common: Always close() TCP sockets on disconnect
In the case of a graceful disconnect of the other peer or if an error
was encountered during recv() or send(), cla_disconnect_handler() is
called on the cla_link. However, the existing handler did not close()
the TCP socket. This led to an exhaustion of file descriptors after many
connections.

This commit adds a TCP-specific disconnect handler which calls close()
on the socket to release the associated resources.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>