cla: Perform an initial connection attempt in TCP-active mode
Single-connection TCP CLAs support an "active" mode, in which they
actively connect to a listening TCP socket. Previously, this connection
was triggered for the first time only upon the first start of a contact.
Thus, if a listening socket is already present when uD3TN starts, the
connection will not be established and, consequently, uD3TN cannot be
reached via the socket until _some_ contact is configured via AAP.

By this change, a single initial connection attempt is made, after which
further connections are only established as long as contacts are
ongoing.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>