test/unit: Add output buffering for Unity tests
This adds a line-based buffer to the putchar functions used for the
Unity-based unit tests. The buffer is flushed either when a newline or
null character is written or when it is full.
This should work around issues with slow processing of the semihosting
output commands used for the debug output (see #52).

Closes: #52

Signed-off-by: Felix Walter <felix.walter@d3tn.com>