CLA: TX: Allow building with CLA TX rate-limiting
This allows us to specify a maximum number of bundles sent per second by
the CLA TX task.

We have to make the bundle transmission less "bursty" in case there is
no additional buffering on the lower layers. For example, if SPP data
received via the TCP socket are directly written to the onboard bus like
on the OPS-SAT.

This smoothes out waiting times by using a constant, pre-calculated
sleep in the TX task if compiled with CLA_TX_RATE_LIMIT defined.

Compile as follows:

    make posix CFLAGS="-DCLA_TX_RATE_LIMIT=<max_bundles_per_second>"

whereas `<max_bundles_per_second>` is an integer number (range of
uint32_t, but not zero).

Signed-off-by: Felix Walter <felix.walter@d3tn.com>