CLA: Make it possible to add an RX read timeout
This adds a compile-time option to add a timeout for reading bytes from
the CLA. If this timeout is exceeded after a CLA read operation, all
parsers will be reset before the received bytes are fed to them.
This feature can be used if it is expected that corrupted or invalid
packets will be received on long-living links, leading to invalid parser
states and no proper processing of valid incoming data.

To enable, compile via:

    make posix CFLAGS="-DCLA_RX_READ_TIMEOUT=<timeout_in_milliseconds>"

Signed-off-by: Felix Walter <felix.walter@d3tn.com>