Rewrite README.md
The README.md has been rewritten from scratch to make uD3TN more
appealing and accessible. It comes with a new structure that aims
to provide a gradual introduction starting from the end user
perspective. Some new content has been added compared to the
previous version:

- Overview figure from the end user perspective
- List of implemented IETF standards / drafts
- CLI usage
- Directory structure and entry point
- List of related tools with a brief description

The mood of the text tries to be as short and precise as possible
with as much information as necessary. Bold and italic highlights
are avoided.

Closes #46 and #61.

Signed-off-by: Georg Alexander Murzik <georg.murzik@d3tn.com>