management_agent: Prevent remote time configuration
We already have a CONFIG_AGENT_REMOTE_CONFIGURATION build option to
enable remote configuration via the config_agent. However, setting the
time via management_agent is allowed in any case. This change applies
the same restrictions to setting the system time.

Note: Setting the time is important for STM32 as this device does not
manage the time by itself like a POSIX-based system.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>