spp_parser: Set error state in case of SPP parsing failure
See #62 - we did not force a parser reset by setting an error state in
case of timecode read errors and unsupported features such as ancillary
data. This properly handles both cases, ensuring that we don't
indefinitely hang in the parsing state and also do not abort due to an
invalid state.

Fixes: #62

Signed-off-by: Felix Walter <felix.walter@d3tn.com>