Move Bundle Status Report's report data init
Adding the bundle source and creation timestamp as well as the fragment
offset and total payload length to the bundle status report report data
was done in the AdministrativeRecord superclass.

This resulted in these values getting appended to the record data of all
administrative records, no matter the type.

This hotfix moves this functionality to the BundleStatusReport subclass,
thereby allowing for the correct initalization of other administrative
record types.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>