Add BIBE PDU definition to bundle.h
uD3TN currently does not accept Bundle-in-Bundle Encapsulation Protocol
Data Units.

This commit is the first step towards adding a BPDU parser, as it adds
the BPDU record type as well as the representation of a BPDU to the
bundle.h file, thereby allowing for the inclusion of the bpdu field in
the bundle_administrative_record struct.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>