Update aap_receive.py to work with BIBE bundles
To allow for receiving and parsing BIBE bundles, the parser implemented
in a previous commit is used. Also in order to differentiate BIBE bundles
from AAP messages, a simple check has been added. This check might need
some refinement if other administrative record types are to be supported
in the future, but it works reliably for BIBE bundles.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>