Fix wrong payload in BIBE adu
So far the BIBE payload of the adu forwarded from the bundle processor
has been the whole administrative record instead of just the BPDU.

This commit remedies this oversight by discarding the administrative
record and just encoding and forwarding the BIBE PDU.

For the encoding to work flawlessly another value has been added to the
bibe_protocol_data_unit struct. The unsigned long payload_length stores
the length of the byte string comprising the encapsulated bundle, so
it's easier to calculate the size of the buffer needed for the cbor
operations.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>