Rework handling of SENDBIBE messages in application agent
Due to fixing the payload of the adu generated in the bundle processor,
as well as a mix-up of AAP message types, parsing a bundle is not re-
quired when handling SENDBIBE messages. Instead, a forward bundle with
the processing flag BUNDLE_FLAG_ADMINISTRATIVE_RECORD has to be created.

This commit modifies the create_forward_bundle and create_bundle methods
of the application_agent.c file to take a parameter "flags" for specify-
ing bundle processing flags that should be applied to the created bundle.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>