mk: Fix build of shared library
Previously, none of the object files were linked into the shared object
libud3tn.so. This adds the appropriate linker flags (--whole-archive) to
add everything to the shared library. -fPIE cannot be used with libs,
thus, is replaced with -fPIC. Finally, the GC options (link-time removal
of unneccessary sections) are only applied for linking the executable.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>