Add chunk read support to AAP parser
As the rx task only provides chunks of incoming messages for parsing,
the AAP parser needed a way to handle this streaming approach.

This commit adds the aap_parser_read function to aap_parser.c and
aap_parser.h, as well as the aap_parser_deinit function, to conform
to the way the CLA RX task handles things.

The aap_parser_read function allows the AAP parser to work with chunks
of AAP messages provided by the RX task, parsing as much as possible
and "requesting" more data when bytes are missing.

The aap_parser_deinit method frees the AAP parser's basedata and clears
the parsed AAP message.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>