Make the remote configuration capability a command line flag
Previously, we had a build-time flag for enabling configuration via
bundles received from other nodes. Thus, changing the flag required a
re-build. As we are depending on it also for our integration tests and
it is often handy in further setups, this change makes it a command
line argument "--allow-remote-config".

As the STM32 platform does not support command line arguments, remote
configuration is enabled by default there via config.h.

See also: !13, #32

Signed-off-by: Felix Walter <felix.walter@d3tn.com>