Add first iteration of the BIBE CLA
This commit adds the first iteration of the BIBE CLA to µD3TN.
For now the BIBE CLA can connect to another µD3TN instance using the
same aap_config.py script the other CLAs also use, register itself as
an application on said other instance and receive messages addressed to
the "bibe" endpoint on the lower layer instance. These messages get sent
to the upper layer (the µD3TN instance the BIBE CLA runs on) via AAP
RECVBIBE messages. The BIBE CLA parses these messages, retrieves the
encapsulated bundle and informs the bundle processor of the incoming
bundle.

So far the reception aspect of the CLA is implemented and working.
Transmission of bundles using the BIBE CLA is not yet supported.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>