Fix minor issues in bibe_over_mtcp_test.py
The default port the script connected to was 4222 instead
of the mtcp port (4224). Also the destinations of the inner and outer
bundles were hardcoded and not changeable via commandline parameters.

This commit fixes those issues by changing the default port and adding
the outer_destination and inner_destination commandline arguments.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>