Add cla_addr parameter to cla_contact_tx_task_command
In order to be able to pass the CLA address to the begin_packet method
of a given CLA, it has to be available within the cla_contact_tx_task.

This commit adds the CLA address to the tx task comman, thus making it
available when calling a given CLA's begin_packet method.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>