Add processing of cla_addr parameter to cla_bibe.c
As the cla_addr parameter contains the EID of the next hop of a BIBE
bundle, it needs to be processed in the cla_bibe.c file.

This commit adds the necessary operations to retrieve the EID from the
cla_addr string. It also adds the necessary function calls for
encoding the BIBE header.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>