Add serialize_pl switch to aap_serializer
In case of a SENDBIBE message generated by the BIBE CLA, the payload
must not be serialized when the rest of the message is serialized.

To make this possible, the serialize_pl parameter was added to the
aap_serializer. If serialize_pl is false, the serializer will skip the
payload and just serialize the rest of the AAP message.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>