Implement BIBE header encoding and transmission
While receiving and parsing BIBE bundles via the BIBE CLA was already
working, transmitting BIBE messages required the encoding of the BPDU,
the administrative record and the enclosing AAP SENDBIBE message
before transmission.

This commit adds this functionality, to allow the BIBE CLA to transmit
and receive bundles as needed.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>