Add BIBE interoperability test with ION
This commit adds an automated interoperability test between µD3TN and
ION. In this test a BIBE bundle is sent to a µD3TN BIBE node which
processes and forwards the encapsulated bundle to an ION instance.
This ION instance discards the encapsulating bundle, encapsulates the
inner bundle again and forwards it to a second µD3TN BIBE node, on which
the receiving bundlesink is running.

This commit also adds "-" to the characters permitted in the CLA string
used to inform the BIBE CLA which node the encapsulating bundle should
be addressed to. This is necessary, as ION uses the "hostname" command
when establishing the administrative endpoint of a node thus creating
an endpoint using the hostname of the machine the node's running on as
EID. This hostname could naturally contain a dash.

Before running the test pyd3tn/bundle7.py has to be modified, as ION uses
administrative record type code 7 to identify BIBE bundles, whereas µD3TN
uses the more recently specified type code 3. In order for the test to
work, the RecordType BIBE_PROTOCOL_DATA_UNIT on line 83 in bundle7.py has
thus to be changed to 7 instead of 3.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>