Fix segementation fault if AR record type unknown
This commit fixes a segmentation fault occurring when the administrative
record parser encountered an unknown administrative record type code.

The cause of the segfault was a misplaced log statement printing the
record type code of the newly parsed administrative record without
checking whether the AR exists at all. Since the AR is NULL in the case
that the AR parser did not recognize the AR typecode, this log statement
tried to access memory it must not access.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>