Address potential memory leak in reports.c
There was an issue in reports.c, that an error at specific points in
the code could lead to already malloc'ed fields not being freed properly.

This commit fixes this issue by introducing the function
free_record_fields() which frees all potentially malloc'ed fields in
the bundle_administrative_records struct before the struct itself
is freed.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>