Refactor BIBE processing on lower layer
With this commit, the lower layer no longer parses the BPDU before sending
it to the upper layer. This also removes the neccessity to re-serialize
the BPDU in the bundle_processor.c file. The lower layer does however
still check whether the BPDU is valid and discards it if it is
malformed in any way.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>