Remove the hal_time dependencies in bundle6 and bundle7
Based on a contribution by @mnitsch, 351e1dae9.

This removes `hal_*` dependencies in the bundle processing components,
with the goal that these parts can be re-used without depending on the
uD3TN core functions and underlying system state. The required variables
are passed as function arguments from the uD3TN core components.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>