CLA: BIBE: Fix memory leak in bibe_begin_packet
In `bibe_begin_packet` another occasion of using `strdup` combined with
`strtok` was found that did not properly `free` the result of `strdup`.
This uses `strchr` instead to get the latter part of the CLA address
string, containing the next-hop EID.

This further checks the provided next-hop EID is valid upon obtaining
a TX queue. By that, the contact manager can properly return an error in
case the EID is invalid.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>