Fix wrong CLA address passed to getaddrinfo()
The cla_sock_addr of the bibe_contact_parameters created in the
launch_connection_management_task() function in cla_bibe.c was set to
the full BIBE-specific CLA address (including the #dtn://xyz.dtn part).
This caused the getaddrinfo() function to fail.

This commit fixes the issue by inserting a null-terminator in place of
the '#' in the CLA address and using this modified cla_sock_addr to init
the bibe_contact_parameters.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>