Assume long-running, high-data-rate contacts have infinite capacity
This allows for adding "infinite capacity" contacts for which the capacity
value is not touched by the router. Note that the bundle size limit of
INT32_MAX is still in place.

Why? -> Previously, we assumed that no more than INT32_MAX bytes can be
transmitted during a single contact, which is unrealistic especially for
long-running/persistent contacts. The forwarding component rewrite will
resolve this once and for all in a clean manner.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>