eid: Support processing `dtn` EIDs without demux having no trailing slash
Previously, we disallowed these EIDs completely, considering them not
valid. Though, this breaks interoperability with ION and seems like an
unnecessary restriction. Local EIDs (the ones uD3TN is initialized with)
still have to contain a trailing slash in any case, the requirements are
just relaxed for processing other EIDs (in the forwarding tables and in
bundle headers).

Signed-off-by: Felix Walter <felix.walter@d3tn.com>