contact_rx_task: Do not drop bundles with unknown source EID schemes
We also allow unknown EID schemes in forwarding, thus, we want to allow
them there as well. The check protects against EID spoofing, but this is
not an issue for EIDs not adhering to either the `dtn` or the `ipn`
scheme as we cannot configure them for the uD3TN instance itself.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>