config_parser: Canonicalize configured node IDs
Our current router implementation searches for the EID returned by
get_node_id(destination). This might differ, e.g., by a slash added
at the end of a `dtn` scheme EID. Thus, we want to ensure that this
value is added to the routing table if we can determine it.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>