tools: Add a "ping" utility
This provides aap_ping.py, a simple tool to send bundles containing just
some information about when they were sent plus a sequence number.
The contained information is printed as soon as the bundles are received
back, similar to the "ping" utility for IP networks. This can be used to
check connectivity and proper operation of an AAP-capable bundle agent
(like uD3TN) and, in combination with a "bundle echo" tool, connectivity
to some arbitrary destination node.

Closes: #79

Signed-off-by: Felix Walter <felix.walter@d3tn.com>