CLA: TCP: Do not attempt to connect if an empty address was specified
We support the case of empty CLA addresses (where the user only
specifies the CLA name). However, this was handled improperly in MTCP,
TCPCLv3, and BIBE, triggering an assertion in
cla_tcp_connect_to_cla_addr. This adds a check ensuring that there is a
non-empty CLA address to which we can attempt to connect.

Fixes: #81

Signed-off-by: Felix Walter <felix.walter@d3tn.com>