CLA: TCP: Always use AF_INET6 while allowing mapped IPv4 addresses
This always creates IPv6 sockets, which are compatible with IPv4
connections via mapped IPv4 addresses. Additionally, the socket option
for only using IPv6 is disabled on systems that set it as default (Note
we can always bind to a specific IPv6 address to force this.)
This way, we get proper dual-stack support.

Fixes: #82

Signed-off-by: Felix Walter <felix.walter@d3tn.com>