Add support for ipn-scheme to BIBE CLA
This commit makes the BIBE CLA run properly on nodes with the ipn-scheme.
Before, the agent needed for the communication between lower and upper
layer always registered with the id "bibe", which does not comply with
the ipn-scheme.

With this commit, the agent checks the EID-scheme of the partner node and
registers either with "bibe", for the dtn-scheme, or 2925 for ipn.

Signed-off-by: Tobias Nöthlich <tobias.noethlich@d3tn.com>