CLA: TCPSPP: Drop SPPs with unknown APID before parsing timecode
We do not need to parse the timecode to drop SPP frames that do not
belong to us. This makes the SPP parser return on every change of its
internal state (as it is called again by the CLA methods, this is a
totally valid mode of operation) and check the APID early in the CLA
methods.

By that we also support receiving SPPs that use different formats to
other APIDs (e.g., without any timecode in case we are expecting a
timecode).

Signed-off-by: Felix Walter <felix.walter@d3tn.com>