CLA: RX: Try to consume remaining buffer after bulk read concludes
In some cases, a bulk read operation may only consume part of the
remaining buffer that was filled by a previous chunk read operation.
If this happens, we just switched back to triggering chunked read, but
were not consuming the remainder of the buffer first. Due to that,
the following call to chunk_read first attempts to fill up the free
second half of the chunk read buffer, delaying reading the remainder of
the data until further data is received.

Fixes: #95
Signed-off-by: Felix Walter <felix.walter@d3tn.com>