CLA: Emit new connection event even if reusing existing connection
If for a CLA that performs an outgoing connection a connection is
already present, it will reuse this. However, it will not tell the
router/BPA that the connection attempt has been successful, which is
however needed to perform actions on pending bundles.

This directly triggers the corresponding event in case a contact was
associated to an existing connection.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>