tools/aap: Use proper logging facilities in tools
This adds a common function to initialize a logger from the Python
`logging` module and uses it in all tools (except aap_test.py, which is
deliberately kept minimal). Each tool chooses in its initialization
routine whether WARN or INFO is the default log level, simply by either
adding 1 to the verbosity parameter or not.

Messages that were previously printed to the terminal using print() are
changed into log statements.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>