router_task: Do not schedule duplicate bundles if fragmentation skipped
The bundle7 fragmenter may return the working bundle when fragmentation
is not needed. This may occur because the router sometimes determines a
longer list of fragments due to conservative estimations of the size of
CBOR ints in the headers of fragments.

If we run into this case, this change makes the router task simply
reduce the total number of fragments to be scheduled and stops the
fragmentation process.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>