router: Respect MBS during long-running contacts
In 0b7f26a2 we introduced a regression due to which fragmentation was
never performed during contacts having a capacity of more than INT32_MAX
bytes (i.e., contacts we consider having "infinite capacity"). This
means that the maximum bundle size (MBS) was not respected during these
contacts, potentially triggering issues on the convergence layer or
below (for example, a MTU may not be respected if the CLA itself does
not segment its packets).

Fixes: #97

Signed-off-by: Felix Walter <felix.walter@d3tn.com>