mk: Also build a static library
This combines the static libraries into one libud3tn.a that can be used
for static linking. It shall be noted that we are using a "thin" archive
(containing symlinks) for maximum compatibility - e.g., the ARM version
of ar does not support conversion and we cannot always assume libtool to
be present. The main objective is to build a library usable for linking
it into uD3TN-NG.

See: https://stackoverflow.com/a/23621751

Signed-off-by: Felix Walter <felix.walter@d3tn.com>