Makefile: Build libraries as well by default
A simple call of `make` will now also build the libraries.

Please note:
We have a quite specific way to re-execute `make` inside the Makefile with
a specific `PLATFORM` set. If we specify multiple targets outside of
this, `make` will execute itself twice concurrently if a `-j` option is
set, leading to race conditions. Thus, we add a new `posix-all` target
that is used to invoke the "inner" `make` process with the proper
`posix` and `posix-lib` targets.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>