router: Remove probabilities and struct routed_bundle
The probability-based routing was never used productively. It implied
that we store all the contacts for which a bundle was scheduled (as
there may be more than one) and keep it, doing reference counting before
we drop the data structure, and so on. This change removes everything
probability-related. The field in the config commands is kept to ensure
backwards compatibility if this is specified.

The removal of struct routed_bundle allows for a major simplification of
all routing-related code.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>