Add flags to enable ASAN and UBSAN in build process
You can now specify `sanitize=yes` to the `make` command, which enables
the address and UB sanitizers. For Clang builds, it enables checks for
unsigned integer ovewrflow and implicit conversions, additionally. Thus,
it is advised to compile as follows:

    make TOOLCHAIN=clang sanitize=yes

There also is the option to pass `sanitize-strict=yes`, which will set
`-fno-sanitize-recover=` such that the program execution aborts in case
the sanitizers detect something.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>