CI: Run unit and integration tests with sanitizers
This enables ASAN and UBSAN in our unit and integration tests.
Another integration test using the thread (data race) sanitizer is added
to detect such issues as well.

The tests are extended such that they wait for uD3TN (all instances if
we launch multiple ones) to terminate, to allow for exit checks of
sanitizers to run. For example, ASAN and the thread sanitizer make use
of this and we need to obtain the exit status code to tell whether the
CI job succeeded or not.

Note that memory leak detection is turned off in the unit tests as they
do not properly deallocate temporary variables, see: #104

aap_send and aap_receive are always invoked with `-v` to get more
information on what happens there in the tests.

Fixes: #55

Signed-off-by: Felix Walter <felix.walter@d3tn.com>