MTCP: Close connection by default
If the other side does not gracefully terminate the TCP connection, we
most likely will not detect it without sending data and running into
timeouts. As the whole idea of the contact scheduling is to define
periods of connectivity, let's assume by default the scheduling is right
and close the connection proactively.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>