hal_io: Forward all logs to stderr and make it line-buffered
This switches all log messages to `stderr`, which is intended for
diagnostic purposes (i.e., matches the intent of our log messages). This
makes it easy to distinguish the log output from actual output (which
the uD3TN daemon does not have, but a program based on libud3tn might).

This also flushes stderr after every log message (i.e., makes it
line-buffered) to prevent delays when outputting the logs.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>