TCPCLv3: Always replace the old connection (make new primary)
Previously TCPCLv3 would operate on a "first come, first served" basis
w.r.t. established connections - if a connection was accepted and there
was already an existing one, the latter was continuously used. However,
as we are not 100% able to determine that a connection is still alive,
this might give us no option to simply replace a broken connection
without waiting for TCPCLv3 to run into a send timeout.

The logic introduced here will always prefer the new connection over any
old one.

Signed-off-by: Felix Walter <felix.walter@d3tn.com>