test/integration: Wait for uD3TN to process node deletion on test end
This ensures the next test's config is not received (via another CLA)
before the deletion has been processed.

Unfortunately, we have no way to determine whether a command has been
processed by uD3TN (there is no operation to query it), so we have to
rely on a simple timeout. This will change with the new forwarding
interface, see the corresponding Milestone.

Fixes: #110

Signed-off-by: Felix Walter <felix.walter@d3tn.com>