CLA: Gracefully handle critical failures
I checked the issues mentioned in #88 - we have two options to handle these
cases; a) continue working but without the broken functionality active
or b) stop the program. The `ASSERT(0)` will do b) in debug builds and
a) in release builds, which is IMO the right way to do it. This commit
just adds a `return` to prevent additional behavior to be triggered in a
release build.

Fixes: #88

Signed-off-by: Felix Walter <felix.walter@d3tn.com>